+++ include/misc/unit_vec.hpp
@@ -23,7 +23,7 @@
 #ifndef _optim_unit_vec_HPP
 #define _optim_unit_vec_HPP
 
-inline arma::vec unit_vec(const int j, const int n);
+inline arma::vec unit_vec(const size_t j, const size_t n);
 
 #include "unit_vec.ipp"
